+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [2.5.0] - 02-25-25
+
+- BREAKING: Removed URL Source (Playlist, Queue, Custom) from SyncPlayer
+- BREAKING: "PlaylistUI" UI prefab has been removed and deleted from "PlayerControls" prefab
+- BREAKING: "Video Source UI" prefab has been added to PlayerControls as replacement for old PlaylistUI
+- Added Source Manager component, which can manage multiple sources (playlists, queues, custom)
+- Added Playlist Queue URL source
+- Added URL Info Resolver component, which can resolve title and author info when loading a YouTube URL 
+- SyncPlayer prefab is updated with a Source Manager added by default with a Queue source
+- Player Controls UI includes an "add to queue" toggle when pressing the [+] change URL button, if a queue is present on an attached source manager
+- Playlist UI rows updated to show track number
+- Playlist UI rows show an "add to queue" button if the backing playlist is associated with a target queue
+
 ## [2.4.15] - 01-07-25
 
 - Added "YouTube Prefer Unity In Editor" option to Video Manager, enabled by default